django/requirements.txt
asgiref==3.7.2
#backports.zoneinfo==0.2.1 # Removido pois gera erro no Python >=3.9
colorama==0.4.6
cx-Oracle==8.3.0
Django==4.2.7
greenlet==3.0.1
logzero==1.7.0
mysqlclient==2.2.0
numpy==1.24.4
pandas==2.0.3
psutil==5.9.7
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
trino
JayDeBeApi # Conexões JDBC - usado para o Connx/Adabas
pip-system-certs # Necessário para validar o certificado durante a conexão com o Trino
oracledb